2.LMStudio_with_PythonCLI/requirements.txt
openai>=1.12.0
rich>=13.7.0
python-dotenv>=1.0.1
lmstudio
